Conflict resolved

See merge request Time-Capture-System/TCS-MobileApp!8

+++ lib/models/task/task.dart
@@ -1,4 +1,4 @@
-import 'package:timecapturesystem/models/task/task_status.dart';
+import 'file:///G:/level_2_project/Git_Lab/TCS-MobileApp/lib/models/task/task_status.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'task.g.dart';
@@ -9,7 +9,7 @@ class Task {
 
   String taskName;
 
-  int estimatedHours;
+  double estimatedHours;
 
   DateTime createdAt;
